Update labmain.c
Generates new platforms and moves the other platforms down by 40 when colliding with it. I have created a list with a couple of possible y positions which the newly generated platforms will get their position from. 

BUG:
-  There is still a bug where the first generated platform remains and the character is not able to hit it. 
- The points are stuck at 8?

The platforms don't scroll - need to implement. 
And maybe add the ground again, which I removed because it was hard to implement.